requirements.txt
configparser
xlrd
requests
numpy
pandas
unidecode
python-levenshtein
